Make code compilable, adding pre-commit and workflows
