+++ .github/workflows/merge.yaml
@@ -1,0 +1,32 @@
+name: On merge
+
+on:
+  pull_request:
+    branches:
+      - master
+    types:
+      - closed
+    paths-ignore:
+      - README.md
+      - .github/**
+      - .*
+
+permissions:
+  contents: write
+
+jobs:
+  tag:
+    name: Tagging
+    runs-on: ubuntu-latest
+    if: github.event.pull_request.merged == true
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Create new tag
+        id: version
+        run: |
+          git pull --tags
+          LAST_TAG=$(git tag --sort=creatordate --sort=v:refname | tail -n1)
+          NEW_TAG=$(echo "$LAST_TAG" | awk '{split(substr($0, 2), a, "."); printf "v%d.%d.%d", a[1], a[2], a[3]+1}')
+          git tag $NEW_TAG
+          git push --tags
