+++ .github/workflows/pr.yaml
@@ -1,0 +1,16 @@
+name: On PR
+
+on:
+  pull_request:
+    branches:
+      - master
+
+jobs:
+  lint:
+    name: Linting
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Run pre-commit
+        uses: pre-commit/action@v2.0.3
