+++ go.mod
@@ -1,0 +1,3 @@
+module github.com/eswdd/go-smee
+
+go 1.17
